refactor: reorganize reservation components and introduce CountdownTimer for better reservation management
